Done for today. Completely functional except for one small bug that came up during testing. "TODO: When the hand has been split on two aces, the aces do not 'buckle down' if necessary. PlayerHand.cs:152"

+++ proj/test/Deck.cs
@@ -3,13 +3,11 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace test
-{
+namespace test {
     /// <summary>
     /// A standard collection of Cards
     /// </summary>
-    class Deck
-    {
+    class Deck {
         /// <summary>
         /// The actual deck - a collection of Cards
         /// </summary>
@@ -17,13 +15,10 @@ namespace test
         /// <summary>
         /// Creates a new Deck and populates it with all 52 standard cards
         /// </summary>
-        public Deck()
-        {
+        public Deck() {
             cards = new List<Card>();
-            foreach (Suit s in Enum.GetValues(typeof(Suit)))
-            {
-                foreach (Rank r in Enum.GetValues(typeof(Rank)))
-                {
+            foreach (Suit s in Enum.GetValues(typeof(Suit))) {
+                foreach (Rank r in Enum.GetValues(typeof(Rank))) {
                     cards.Add(new Card(r, s, this));
                 }
             }
@@ -33,8 +28,7 @@ namespace test
         /// Returns true if there are no cards left in the Deck, false otherwise
         /// </summary>
         public bool IsEmpty {
-            get
-            {
+            get {
                 return Count == 0;
             }
         }
@@ -42,19 +36,16 @@ namespace test
         /// <summary>
         /// The number of cards currently in the Deck
         /// </summary>
-        public int Count
-        {
-            get
-            {
-                return cards.Count;
+        public uint Count {
+            get {
+                return (uint)cards.Count;
             }
         }
 
         /// <summary>
         /// Returns a random card from this Deck, removing it internally
         /// </summary>
-        public Card Draw()
-        {
+        public Card Draw() {
             if (IsEmpty)
                 return null;
             Card r = cards[(new Random()).Next(cards.Count)];
@@ -65,8 +56,7 @@ namespace test
         /// <summary>
         /// Returns a String representation of this Deck
         /// </summary>
-        public override string ToString()
-        {
+        public override string ToString() {
             return cards.Count + " " + cards.ToString();
         }
